Backlog
The backlog performs two essential functions:

1.  It provides a way for a user to float an idea before committing
to a detailed analysis.  It is, instead, a way for a user to get
approval on an approach or design.

2.  It is a parking lot for specifications that are orphaned.  If a
spec is not going to be implemented by the assigned user, it will be
collected into the backlog, and be available for other developers to
implement.

Co-Authored-By: Steve Martinelli <stevemar@ca.ibm.com>
Change-Id: If546724fd535db7753a372389c3f90f3b060d9bc