Split up assignments, making role-assignments pluggable.
Split the current "assignments" components into the base entities
(i.e. domain, project & role) and the actual assignment model.
Make the (real) assignment part of this pluggable, so as to allow
alternative models to be introduced.

In addition, provide a base to more easily rationalize the
duplicate methods of extracting assignments.

Change-Id: I46b8298559ff6040591a9fbd02a69f99ed859af9
Implements: bp pluggable-assignment