Replace the concept of extensions in Keystone.
The existing concept of optionally wsgi-loaded extensions has been
shown to cause complexity and confusion for clients of a service
that supports such extensions.

This spec proposes we remove this concept in favor of code
either being in-tree (and classed as stable or experimental) or
out-of-tree, in which case we provide no guarantees or support.

Change-Id: I06b0ecc3d02a2a1b7ad5cf803009574ed2376783