Do not add new 'db' command and subcommands for it
At the moment of writing of the spec, there was an intention to provide
a uniform way for interacting with various migration managers. Since
that time, Alembic went far forward in its features and added some new
commands, which are not present in sqlalchemy-migrate, making the interface
irrelevant.

Remove mentions of the 'db' command and leave existing db_sync and
db_version commands.

Change-Id: I7ada206c3a8d8c0bd5625cc563e799a98c3ce23b