Be consistent in how we give error codes in the Identity spec
Throughout the spec, we use various ways of given the HTTP error
codes, sometimes quoted, sometimes not, sometimes with the "HTTP"
moniker, sometimes not.

This patch makes this consistent, using the form:

HTTP ``400 Bad Request``

No actual changes to the error code being returned are being
made, this is just about formatting the document.

Change-Id: Icc21ebd1262ae889c7dcc4ee43ad203d52bcab0d