Improve example of project acting as a domain
The current project example in the project entity section was,
in fact, invalid - it showed a project acting as a domain that
was not at the top of the tree. This patch corrects this and adds
correct examples of a project acting as a domain and a top level
regular project.

Change-Id: I2b0519f3c0ea274898ffd2ed6cad24c6b4352940