Simplify manage-migration spec by introducing database triggers
By introducing database triggers to this specification, we can vastly
simplify the overall implementation and operator experience and skip
straight to zero downtime. There's no need for the operator to manage
any metadata in the database about the state of the migration, beyond
what sqlalchemy-migrate will do for us, and there's no need to write
versionedobjects for each possible object state during a live migration.
Hopefully that will all translate to fewer bugs.

On the downside, it means that we have to manually write triggers, which
may be need to be specifically tailed for each of our supported
databases (MySQL, PostgreSQL, and SQLite).

Change-Id: I38c02a3a3388ed57032f01e8f522ce53230231f8