OpenID Connect improved support
OpenID Connect is supported in Keystone by leveraging the Apache
mod_auth_oidc module and the Keystone Federation plugin. OpenID Connect
works fine when accessing OpenStack thorugh the dashboard, but it
requires additional configuration steps to make it work when using the
OpenStack CLI tools. This spec aims at improving the support, so
that the same outcome is obtained, regardless of the way the user
accesses the cloud.

Implements blueprint: improved-oidc-support

Change-Id: I97f7a34d398ba673d7733fb2aaa490ebc9298afd