Add capabilities to application credentials
This spec describes a capability list extension for
application credentials that allows their creator to restrict
their usage to a white list of capabilities, each of which
consists of HTTP methods and permitted URL paths. Any request
using the application credential must match an entry in this
list. All other requests are rejected.

Change-Id: Iffc853971e909bd3ca8829b410a7014929c8b7e6