Update shadow mapping details about domains.
The shadow mapping specification originally implied that the domain
id would come from an attribute in the mapping. Through the process
of implementing shadow mapping, and the other benefits of having a
one-to-one mapping between Identity Providers and domains, it became
apparent that the shadow mapping can just pull the domain information
from the Identity Provider in the mapping process instead.

bp shadow-mapping

Change-Id: I88c0c38cdd78bfa570463f4b62b659c436b73419