Remove the fernet key store spec from backlog
This was discussed in our weekly meeting:

  http://eavesdrop.openstack.org/meetings/keystone/2017/keystone.2017-02-28-18.00.log.html#l-94

We realize the need for alternate storage methods for sensitive
information, especially fernet keys. Having a pluggable backend
specifically for fernet keys was slightly specific and posed some
implementation challenges. This commit removes the spec from the
backlog such that we aren't misleading new developers or people who
reference the backlog frequently to get an idea of project direction.

Change-Id: I703c93be43f759102d59f3dc3b0e1737d46e54b4