Clarify bits of the alembic backlogged spec
We discussed the need for this spec in the keystone meeting after the
Pike PTG:

  http://eavesdrop.openstack.org/meetings/keystone/2017/keystone.2017-02-28-18.00.log.html#l-203

The requirement for implementing this specification has a hard
dependency on supporting the OpenStack sqlalchemy-migrate fork. If
that support ever goes away, we'll *have* to move forward with this
approach.

This commit updates the spec to reflect those changes such that new
developers and people who reference our backlog for project direction
are aware and informed.

Change-Id: Ic44d28943512f9a26452bb2b98c62f656d2ea859