Remove centralized policies fetch cache spec
This was discussed in our week keystone after the Pike PTG:

  http://eavesdrop.openstack.org/meetings/keystone/2017/keystone.2017-02-28-18.00.log.html

It was determined that with the current direction of policy across
OpenStack projects, that this spec is no longer needed. Depending on
the resulting discussions, we can repropose an update and accurate
spec. Until that happens, we should remove this from the backlog so
that this doesn't mislead new developers or people who use our
backlog to gauge project direction.

Change-Id: Ic8d54b13b7f9a2d5fa4314ea578d6786be08f0f6