Application Credentials for application authn
Implement Applicataion Credentials as a new credential type in order to
improve security by limiting access and not embedding user account
credentials in configuration files. Application Credentials are intended
to empower consumers of clouds to have a mechanism to safely use the API
without compromising their credentials, and to allow cloud-native
applications that need to make API calls the ability to do so without
needing to have the complete account credentials embedded.

Partially-implements: blueprint application-credentials
Co-Authored-By: Jamie Lennox <jamielennox@gmail.com>
Co-Authored-By: Monty Taylor <mordred@inaugust.com>
Co-Authored-By: Colleen Murphy <colleen.murphy@suse.com>
Change-Id: I58a345513dadbacd4166f7d4a4b8d11775ccc736