Clarify backlog instructions and add ideas dir
One of the outcomes of the Pike retrospective to clarification
of the purpose of backlog. It was also mentioned that it would be
useful to have an `ideas` directory that allow people to propose
ideas that need to be freshed out and discussed. The backlog somewhat
filled this purpose but is going to be more specific in the future.

This commit clears up the direction of each location and what purpose
they fill.

Change-Id: Ie07f11e67aef1e03311c430963938fe079bf0fa6