Define a set of basic default roles
With the recent work to keystone and oslo.policy, we should be able to
offer tooling to project development teams so they can start evolving
their policies. Before we start changing things, we should come to
consensus on a set of defaults we should offer out of the box.

Moving towards a set of known defaults will make maintenance for
operators much easier. It will also build the foundation for a more
robust RBAC system that is better at modeling complex
organization, ultimately being more useful in the real-world.

Change-Id: Ia6ddf64b4483a73ab79c86d5d794cce561aa19e0
Co-authored-By: Lance Bragstad <lbragstad@gmail.com>
Co-Authored-By: Jamie Lennox <jamielennox@gmail.com>