Fix broken link to Stein roadmap
The link to the Stein roadmap was populated before we actually went
through the schedule for Stein. We actually used a copy of a previous
roadmap to build the Stein roadmap, making it easier to manage carry
over items. As a result, the original link isn't useful and was
abandoned.

This commit updates the link to point to the correct roadmap.

Change-Id: Ib02bd11b3604c366db873c0f74d739dd04d322e2