Update template with guidance for bugs over bp
Change the template and provide guidance for using a LP bug instead
of the blueprints. This is due to the differences in blueprints and
bugs. Blueprints do not have comments and the entire body is mutable.
With the entire content being mutable (and having no real history
functionality) anyone can change the blueprint and lose all tracking
of patches proposed for the work.

When a bug is used instead of a blueprint, each patch proposal and
merge generates an immutable comment. Short of deleting the bug, the
tracking of the work that has been done cannot be lost.

Bugs can also be closed with a ``closes-bug`` line in the commit
messages where specs take extra human intervention to mark as complete.

Change-Id: Id575258f98baddc7e4fe1ebfe5c00e9ea3a87e9f