Move the request-helpers spec for keystonemiddleware to attic
Move the request-helpers backlog spec for keystonemiddleware to the
attic. At the Denver PTG (2019) we discussed this spec. We are in a
very different space from where we were at the time of proposal, and
if there is a desire to revisit this specific specification it can
be brought back from the attic.

Change-Id: I3e1ab025bb998b14c0a71854b9109d9f29b25ee9