Repropose Expiring Group Membership for Ussuri
Add federated users to the groups that they receive from the mapping rules.
This membership is only carried by the token and not persisted in the
database. The membership expires, but can be renewed when the user
authenticates with the same group.

Previously approved for Train, fell into backlog, reproposing for Ussuri.

Change-Id: Ie133c14ffba5e4189265920759bfb5e1391f1189
Partial-Bug: 1809116