Disable auto-discovery for setuptools and update python testing
With setuptools release 61.0.0 docs build started to fail:
error: Multiple top-level packages discovered in a flat-layout:
['specs', 'attic', 'superseded'].

This bug is mentioned in setuptools issue 3197 [0], and the suggested
workaround is to disable auto-discovery by adding 'py_modules=[]' in
setup.py.

Also use recent python versions because the old ones are no longer available.

These 2 unrelated changes need to be merged together in order to unblock
the gate.

[0] https://github.com/pypa/setuptools/issues/3197

Change-Id: Iddc30b9521b61d9083c2b1f6e8a6707196ea0a57